2010-01-14 11:31 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/vm/set.c   * src/debug/dbgentry.c   * src/common/hbgete.c   * src/common/hbffind.c   * src/common/hbwince.c   * src/common/hbstr.c   * src/common/hbmem.c   * src/common/strwild.c   * src/common/hbhash.c   * src/rtl/lennum.c   * src/rtl/strxor.c   * src/rtl/strcase.c   * src/rtl/hbzlibgz.c   * src/rtl/trim.c   * src/rtl/base64c.c   * src/rtl/binnum.c   * src/rtl/hbtoken.c   * src/rtl/ampm.c   * src/rtl/descend.c   * src/rtl/strpeek.c   * src/rtl/replic.c   * src/rtl/padr.c   * src/rtl/hbstrfmt.c   * src/rtl/hbntos.c   * src/rtl/hbadler.c   * src/rtl/at.c   * src/rtl/fserr.c   * src/rtl/hbcrc.c   * src/rtl/hbmd5.c   * src/rtl/hbascii.c   * src/rtl/padc.c   * src/rtl/strzero.c   * src/rtl/hbregexc.c   * src/rtl/strtran.c   * src/rtl/hbsocket.c   * src/rtl/hbinet.c   * src/rtl/strtoexp.c   * src/rtl/gtapi.c   * src/rtl/filesys.c   * src/rtl/padl.c   * src/rtl/hbregex.c   * src/rtl/saverest.c   * src/rtl/right.c   * src/rtl/hbbffnc.c   * src/rtl/cdpapi.c   * src/rtl/hbi18n1.c   * src/rtl/gtclip.c   * src/rtl/mlcfunc.c   * src/rtl/left.c   * src/rtl/xsavescr.c   * src/rtl/hardcr.c   * src/rtl/setkey.c   * src/rtl/hbzlib.c   * src/rtl/strc.c   * src/rtl/stuff.c   * src/rtl/hbhex.c   * src/rtl/trace.c   * src/rtl/samples.c   * src/rtl/inkeyapi.c   * src/rtl/inkey.c   * src/rtl/mtran.c   * src/rtl/gete.c   * src/rtl/mouse53.c   * src/rtl/colorind.c   * src/rtl/ati.c   * src/rtl/soundex.c   * src/rtl/memofile.c   * src/rtl/valtostr.c   * src/rtl/rat.c   * include/hbapicdp.h   * include/hbgtcore.h   * include/hbapifs.h   * include/hbapi.h   * include/hbchksum.h   * include/hbapigt.h   * include/hbhash.h   * contrib/xhb/hboutdbg.c   * contrib/xhb/hbxml.c   * contrib/xhb/hbcrypt.c   * contrib/xhb/dbf2txt.c   * contrib/xhb/xstrdel.c   * contrib/xhb/xhbsave.c   * contrib/xhb/hbcomprs.c   * contrib/xhb/datesxhb.c   * contrib/xhb/xhbat.c   * contrib/xhb/xhbmsgs.c   * contrib/xhb/xhbarr.c   * contrib/xhb/txtline.c   * contrib/xhb/xhbtrim.c   * contrib/xhb/hbserv.c   * contrib/xhb/cstructc.c   * contrib/xhb/xhbfunc.c     * ULONG -> HB_SIZE
  * src/rtl/langapi.c
  * include/hbapilng.h
    * ULONG -> int
    + Marked hb_langDGetErrorDesc() with HB_LEGACY_LEVEL3.

  * src/common/hbprintf.c
    ! ULONG -> int

  * src/debug/dbgentry.c
    * Formatting.

+++ harbour/src/common/hbprintf.c
@@ -1304,7 +1304,7 @@ int hb_vsnprintf( char * buffer, size_t bufsize, const char * format, va_list ap
 /* NOTE: The full size of the buffer is expected as nSize. [vszakats] */
 int hb_vsnprintf( char * buffer, size_t nSize, const char * format, va_list arglist )
 {
-   ULONG result;
+   int result;
 
 #if defined( __DJGPP__ ) && ( __DJGPP__ < 2 || ( __DJGPP__ == 2 && __DJGPP_MINOR__ <= 3 ) )
    /* Use vsprintf() for DJGPP <= 2.03.
